Add Sale Report API

+++ QuanLyQuanCafe.Server/Models/DTOs/MonthSalaryStatisticDTO.cs
@@ -6,5 +6,5 @@ public class MonthSalaryStatisticDTO
 {
     public string Month { get; set; }
     public int TotalHours { get; set; }
-    public decimal TotalSalaryPayed { get; set; }
+    public double TotalSalaryPayed { get; set; }
 }
